refactor: query-string 선언 service.ts로 이동
