+++ src/components/my/AllProjects.tsx
@@ -9,8 +9,7 @@ import useRequest from "@/hooks/use-request"
 import useInput, {type TInputChangeEvent} from "@/hooks/use-input"
 import usePagination from "@/hooks/use-pagination"
 import {calculateRemainingDay} from "@/utils/formatter"
-import {ROUTE} from "@/constants/service"
-import {OPEN, ALL_PROJECTS} from "@/constants/query-string"
+import {OPEN, ALL_PROJECTS, ROUTE} from "@/constants/service"
 
 const BottomSheet = dynamic(() => import("../BottomSheet"), {
   loading: () => <></>,
