--- src/constants/query-string.ts
@@ -1,5 +1,0 @@
-export const OPEN = "open"
-
-export const ALL_PROJECTS = "all-projects"
-
-export const ALL_RECEIVED_ROLLING_PAPERS = "all-received-rolling-papers"
